refactor(onboarding): consolidate navigation helpers into OnboardingNavigationUtils
