--- src/libs/Navigation/helpers/dismissOnboardingModalBeforeExit/index.native.ts
@@ -1,7 +1,0 @@
-import Navigation from '@libs/Navigation/Navigation';
-
-function dismissOnboardingModalBeforeExit() {
-    Navigation.dismissModal();
-}
-
-export default dismissOnboardingModalBeforeExit;
