--- src/libs/Navigation/helpers/dismissOnboardingModalBeforeExit/index.ts
@@ -1,3 +1,0 @@
-function dismissOnboardingModalBeforeExit() {}
-
-export default dismissOnboardingModalBeforeExit;
