--- src/libs/Navigation/helpers/getOnboardingExitNavigationOptions/index.native.ts
@@ -1,7 +1,0 @@
-import type {LinkToOptions} from '@libs/Navigation/helpers/linkTo/types';
-
-function getOnboardingExitNavigationOptions(): LinkToOptions | undefined {
-    return undefined;
-}
-
-export default getOnboardingExitNavigationOptions;
