--- src/libs/Navigation/helpers/getOnboardingExitNavigationOptions/index.ts
@@ -1,7 +1,0 @@
-import type {LinkToOptions} from '@libs/Navigation/helpers/linkTo/types';
-
-function getOnboardingExitNavigationOptions(): LinkToOptions | undefined {
-    return {forceReplace: true};
-}
-
-export default getOnboardingExitNavigationOptions;
