--- src/libs/Navigation/helpers/resetOnboardingStackToRoot/index.native.ts
@@ -1,3 +1,0 @@
-function resetOnboardingStackToRoot() {}
-
-export default resetOnboardingStackToRoot;
