+++ src/libs/navigateAfterOnboarding.ts
@@ -13,9 +13,8 @@ import Onyx from 'react-native-onyx';
 import {setDisableDismissOnEscape} from './actions/Modal';
 import SidePanelActions from './actions/SidePanel';
 import {setOnboardingRHPVariant} from './actions/Welcome';
-import dismissOnboardingModalBeforeExit from './Navigation/helpers/dismissOnboardingModalBeforeExit';
-import getOnboardingExitNavigationOptions from './Navigation/helpers/getOnboardingExitNavigationOptions';
 import isReportTopmostSplitNavigator from './Navigation/helpers/isReportTopmostSplitNavigator';
+import {dismissOnboardingModalBeforeExit, getOnboardingExitNavigationOptions} from './Navigation/helpers/OnboardingNavigationUtils';
 import shouldOpenOnAdminRoom from './Navigation/helpers/shouldOpenOnAdminRoom';
 import Navigation from './Navigation/Navigation';
 import {findLastAccessedReport, isConciergeChatReport, isSelfDM} from './ReportUtils';
